backend_function/requirements.txt
boto3==1.17.29
botocore==1.20.29
requests==2.22.0
PyJWT==2.0.1

